requirements.txt
# --- Flux Audio & VAD (Local - Compatible Python 3.13) ---
# pip install torch torchaudio --extra-index-url https://download.pytorch.org/whl/cu121
openai
python-dotenv
sounddevice
numpy==1.26.4
torch
packaging
# --- Vecteur & Maths ---
chromadb
# --- HTTP Client (Obsidian Bridge) ---
httpx
urllib3
# --- Synthèse Vocale (P3) ---
pygame==2.5.2
edge-tts
asyncio
#--- Visualisation des tableaux d'exécution ---
rich
fastapi
uvicorn
Jinja2
python-multipart
#--- pydantic-settings ---
pydantic-settings
PyYAML
pydantic>=2.0.0